Corrigindo Insert Usuario

+++ app/src/main/java/com/example/refugiohotelsapp/Entidades/Reserva.java
@@ -115,7 +115,7 @@ import java.util.Date;
         this.reservaAtiva = reservaAtiva;
     }
 
-    public Long getIdUsuario() {
+    public Integer getIdUsuario() {
         return usuario.getIdUsuario();
     }
